model/requirements.txt
tensorflow>=2.13.0
numpy>=1.23.0
Pillow>=9.0.0
